feat : POST UPDATE & DELETE 시 Authorization 검증

+++ src/main/java/com/kt/board/service/PostServiceImpl.java
@@ -2,6 +2,8 @@ package com.kt.board.service;
 
 import com.kt.board.config.RedisConfig;
 import com.kt.board.redis.TokenStore;
+
+import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,6 +26,7 @@ public class PostServiceImpl implements PostService {
 	private final BoardRepository boardRepository;
 	private final UserRepository userRepository;
     private final TokenStore tokenStore;
+		private final StringRedisTemplate redisTemplate;
 
 	@Transactional
 	@Override
@@ -49,14 +52,23 @@ public class PostServiceImpl implements PostService {
 	public void update(Long postId, PostUpdateRequest request, String authorization){
 		PostEntity post = postRepository.findById(postId)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시글입니다."));
-        post.update(request.title(), request.content(), request.disclosureType());
+		Long userId = post.getCreatedBy().getId();
+		System.out.println(authorization);
+		System.out.println(userId);
+		if (!tokenStore.isValidToken(userId,authorization)) throw new IllegalArgumentException("일치하지 않는 사용자입니다.");
+
+		post.update(request.title(), request.content(), request.disclosureType());
 	}
 
 	@Transactional
 	@Override
-	public void remove(Long postId){
+	public void remove(Long postId, String authorization){
 		PostEntity postEntity = postRepository.findById(postId)
 			.orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시글입니다."));
+
+		Long userId = postEntity.getCreatedBy().getId();
+		if (!tokenStore.isValidToken(userId,authorization))  throw new IllegalArgumentException("일치하지 않는 사용자입니다.");
+
 		postRepository.delete(postEntity); // soft delete
 	}
 }
